Install postgresql on share
This allows the `:share` VM to be used as a import location for the
database dumps/backups, without requiring shutting down the main
appliance DB to validate it.

The 3.4 repository for alpine is used here, since it has the only
location of the `5.6.x` version of postgresql available, and we want the
version numbers to match up.  Granted, it is 5.6.12 and the
gaprindashvili version of postgresql is older than that, but only a
difference in patch so it should be good enough for testing.


(transferred from https://gist.github.com/NickLaMuro/8438015363d90a2d2456be650a2b9bc6/195817fd2f73a8bcd0e15d3b5726c43bfba0d052)