Make sure share has base config to connect to
Previously, we had always tested the DB backups prior to the DB dumps,
which part of the backup validation is we have to reload the entire
thing, config and all.

There is a large portion of the DB helper that is in charge of making
sure that the new backup can be connected to also handles making sure
the new config that is loaded from the backup conforms to the VM setup
we have.  This coincidently allowed the DB Dump tasks to pass when the
DB backup tests had already run.

This makes it so that if the DB dump tasks are run first, no issues with
connecting will occur.


(transferred from https://gist.github.com/NickLaMuro/8438015363d90a2d2456be650a2b9bc6/2cfeaf88a568070e2bf2f4d7a4f296ce52fb25e9)