[Vagrantfile] Remove debug for pub key
When this repo was originally being put together, I was using this debug
statement just to determine the result of the if statement.  At the
time, the case was reversed, so the output was meant to deploy when the
key already existed, despite the fact that the conditional block the
`puts` is a part of currently creates the non-existent key.

This debug statement is no longer valid, or needed in its current form.


(transferred from https://gist.github.com/NickLaMuro/8438015363d90a2d2456be650a2b9bc6/91a85384c6fe89bef79ab0961d32883322a8c54d)