[NEW] Support testing backups
Adds a bunch of support code for "properly" validating backups.

The `appliance_secondary_db.rb` file is meant to be a helper lib for
providing a consistent interface for managing a secondary DB via
`pgctl`.

`appliance_rake_restore_patches.rb` allows for that interface to patch
`LinuxAdmin::Service` and `PostgresAdmin` so that they instead use the
"secondary DB" instead of the default one.

We remove any code to now do this in the Vagrantfile as the above can be
used in more places (and doesn't need to be on the host machine).

* * *

The above allows the code for MIQ to mostly run as it does to backup and
restore a database backup, but doesn't blow the reference database away
in the process (which is very important).

Other code changes for existing files support these changes and allow
them to be either used from the testing code directly, or that they are
pointing to the correct database instance when running tests or backups.


(transferred from https://gist.github.com/NickLaMuro/8438015363d90a2d2456be650a2b9bc6/d6898c404e7f16c355ff395aae6799e46e80c2ae)