[appliance_console_backups] Use master appliance
Since a patch was required to fix an issue with hammer:

https://github.com/ManageIQ/manageiq/pull/18745

Certain specs do not work with the rake tasks, and causes a decent
amount of errors.

As a result, the simplest way forward was to just simply switch to using
a master appliance for the time being.  However, the downside to this is
that "master" isn't an available option to download from vagrantup.com:

https://app.vagrantup.com/manageiq

So downloading and installing an appliance from:

http://releases.manageiq.org/

is required for this to work.  A more pragmatic solution in the future
might be to allow setting a custom box, but defaulting to a box type
that available from vagrantup.com.

Steps to download/install manageiq/master:
------------------------------------------

Since the `manageiq/master` appliance isn't available as valid release
from https://app.vagrantup.com/manageiq/ , you can use the following
script to download a versioned copy of it:

https://github.com/NickLaMuro/miq_tools/pull/13

And add it to vagrant by running the following:

    $ ./miq_vagrant_master/cli --version 20190629