[Rakefile] Use __dir__ for read_vm_state
Vagrant names it's VMs based on the `File.basename` for the directory
the Vagrantfile is in, as well as the name given for the VM in
Vagrantfile configuration.

Allows matching the new project name, as well as allowing users to have
change the directory they clone this project into as well, and this will
still function.