Convert to minitest
Instead of using a custom test runner, minitest is included in almost
every version of Ruby, so use that instead.

The directory structure changed heavily as a result, but was mostly
moved from tests/ -> test/

A Rakefile was also included at the root of the test dir that can be
used for executing tests on the appliance.  This also makes use of the
`rake/clean` lib to handle cleaning instead of writing something custom.