[ApplianceConsoleRunner] Update for less prompts
When updating the DatabaseAdmin tasks for the appliance console, there
will be less options available to input.

No longer will it ask for a type, as it will always assume you are
writing to a "local file" (whether that file is a mounted NFS/SMB/etc.
share or just a local disk will not be handled here).

As such, ApplianceConsoleRunner no longer needs to match against the
IO_MENU_OPTION_REGEXPS as it did previously, so they have been removed.