[test/Rakefile] Fix test task
Running all of the tests was not working since the `Dir.pwd` was usually
run from the login dir, as the "outer Rakefile" (`./Rakefile`) was just
passing the `--rakefile` flag.

This mean that "./**/*_test.rb" was trying to find test dirs from
`/home/vagrant` instead of from `/vagrant/test`, which didn't work.