[test/{nfs|smb}*.rb] Update for self mounting
Moving forward, we will not be supporting the automatic connection of
file storage backends, so this will now need to be managed by the user
themselves.

By association, this means the tests themselves also need to manage the
connection.  To ensure the connection is working properly, we connect
once for creating the backup, and a separate time (in a new location)
when validating the DB.  This helps ensure that we are not just writing
randomly to the filesystem, but the mounted network drive as expected.